Move docs components around to match create-react-app
