+++ .corkboard/src/components/Checkerboard.js
