+++ .corkboard/src/components/Example.js
@@ -1,13 +1,13 @@
 import React from 'react';
-import Box from '../src/Box/Box';
-import Text from '../src/Text/Text';
+import Box from '../../../src/Box/Box';
+import Text from '../../../src/Text/Text';
 import Checkerboard from './Checkerboard';
-import Link from '../src/Link/Link';
-import Icon from '../src/Icon/Icon';
-import Checkbox from '../src/Checkbox/Checkbox';
-import Column from '../src/Column/Column';
-import Mask from '../src/Mask/Mask';
-import Heading from '../src/Heading/Heading';
+import Link from '../../../src/Link/Link';
+import Icon from '../../../src/Icon/Icon';
+import Checkbox from '../../../src/Checkbox/Checkbox';
+import Column from '../../../src/Column/Column';
+import Mask from '../../../src/Mask/Mask';
+import Heading from '../../../src/Heading/Heading';
 
 import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live';
