+++ .corkboard/src/components/Navigation.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Text from '../src/Text/Text';
-import Box from '../src/Box/Box';
-import SelectList from '../src/SelectList/SelectList';
-import Link from '../src/Link/Link';
-import Icon from '../src/Icon/Icon';
-import Heading from '../src/Heading/Heading';
+import Text from '../../../src/Text/Text';
+import Box from '../../../src/Box/Box';
+import SelectList from '../../../src/SelectList/SelectList';
+import Link from '../../../src/Link/Link';
+import Icon from '../../../src/Icon/Icon';
+import Heading from '../../../src/Heading/Heading';
 
 const isLeftClickEvent = event => event.button === 0;
 const isModifiedEvent = event =>
