+++ .corkboard/src/index.js
@@ -1,12 +1,12 @@
 import 'corkboard/init';
 import React from 'react';
 import { Router, hashHistory, Route, IndexRoute } from 'react-router';
-import App from './App';
-import CardPage from './CardPage';
+import App from './components/App';
+import CardPage from './components/CardPage';
 import { render } from 'react-dom';
 import { getCards } from 'corkboard/init';
 
-const requireCard = require.context('../src', true, /[\-\.]doc\.js$/);
+const requireCard = require.context('../../src', true, /[\-\.]doc\.js$/);
 const paths = requireCard.keys();
 paths.sort((a, b) => a.localeCompare(b));
 paths.forEach(requireCard);
