+++ src/Card/Card.doc.js
@@ -7,7 +7,7 @@ import Button from '../Button/Button';
 import Card from './Card';
 import Link from '../Link/Link';
 import Text from '../Text/Text';
-import { ns, card, md, PropTable, Example } from '../../.corkboard/cards';
+import { ns, card, md, PropTable, Example } from '../../.corkboard/src/cards';
 
 ns(
   'Card',
