+++ src/Checkbox/Checkbox.doc.js
@@ -11,7 +11,7 @@ import {
   PropTable,
   Example,
   Combination,
-} from '../../.corkboard/cards';
+} from '../../.corkboard/src/cards';
 
 ns(
   'Checkbox',
