+++ src/GroupAvatar/GroupAvatar.doc.js
@@ -8,7 +8,7 @@ import {
   PropTable,
   Example,
   Combination,
-} from '../../.corkboard/cards';
+} from '../../.corkboard/src/cards';
 
 ns(
   'Group Avatar',
