+++ src/SearchField/SearchField.doc.js
@@ -2,7 +2,14 @@
 
 import * as React from 'react';
 import SearchField from './SearchField';
-import { ns, md, card, PropTable, Example } from '../../.corkboard/cards';
+import {
+  ns,
+  card,
+  PropTable,
+  Example,
+  StateRecorder,
+} from '../../.corkboard/src/cards';
+import Box from '../Box/Box';
 import Icon from '../Icon/Icon';
 import IconButton from '../IconButton/IconButton';
